+++ packages/orq-rc/src/orq_ai_sdk/models/createpromptop.py
@@ -1961,6 +1961,7 @@ CreatePromptProvider = Literal[
     "minimax",
     "xai",
     "alibaba",
+    "tensorix",
     "slack",
     "orq",
 ]