+++ packages/orq-rc/src/orq_ai_sdk/models/deploymentinvokeop.py
@@ -87,6 +87,7 @@ Provider = Literal[
     "minimax",
     "xai",
     "alibaba",
+    "tensorix",
     "slack",
     "orq",
 ]