+++ packages/orq-rc/src/orq_ai_sdk/models/deploymentsop.py
@@ -632,6 +632,7 @@ DeploymentsProvider = Literal[
     "minimax",
     "xai",
     "alibaba",
+    "tensorix",
     "slack",
     "orq",
 ]