+++ packages/orq-rc/src/orq_ai_sdk/models/deploymentstreamop.py
@@ -2213,6 +2213,7 @@ DeploymentStreamProvider = Literal[
     "minimax",
     "xai",
     "alibaba",
+    "tensorix",
     "slack",
     "orq",
 ]