+++ packages/orq-rc/src/orq_ai_sdk/models/getallpromptsop.py
@@ -536,6 +536,7 @@ GetAllPromptsProvider = Literal[
     "minimax",
     "xai",
     "alibaba",
+    "tensorix",
     "slack",
     "orq",
 ]