+++ packages/orq-rc/src/orq_ai_sdk/models/getonepromptop.py
@@ -499,6 +499,7 @@ GetOnePromptProvider = Literal[
     "minimax",
     "xai",
     "alibaba",
+    "tensorix",
     "slack",
     "orq",
 ]