+++ packages/orq-rc/src/orq_ai_sdk/models/getpromptversionop.py
@@ -535,6 +535,7 @@ GetPromptVersionProvider = Literal[
     "minimax",
     "xai",
     "alibaba",
+    "tensorix",
     "slack",
     "orq",
 ]