+++ packages/orq-rc/src/orq_ai_sdk/models/listpromptversionsop.py
@@ -545,6 +545,7 @@ ListPromptVersionsProvider = Literal[
     "minimax",
     "xai",
     "alibaba",
+    "tensorix",
     "slack",
     "orq",
 ]