+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KN1EM1GTNPKS38F7E9JMWCN4"
+        "01KN1QEWC92MSPBB2C42K1RKQ5"
     )
     r"""The unique identifier of the data source"""