+++ packages/orq-rc/src/orq_ai_sdk/models/updatepromptop.py
@@ -2059,6 +2059,7 @@ UpdatePromptProvider = Literal[
     "minimax",
     "xai",
     "alibaba",
+    "tensorix",
     "slack",
     "orq",
 ]