+++ packages/orq-rc/RELEASES.md
@@ -6739,3 +6739,13 @@ Based on:
 - [python v4.7.0-rc.33] packages/orq-rc
 ### Releases
 - [PyPI v4.7.0-rc.33] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.33 - packages/orq-rc
+
+## 2026-03-31 10:39:29
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.0 (2.879.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.7.0-rc.34] packages/orq-rc
+### Releases
+- [PyPI v4.7.0-rc.34] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.34 - packages/orq-rc